% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.dca}
\alias{as_tibble.dca}
\title{Export DCA tibble}
\usage{
\method{as_tibble}{dca}(x, ...)
}
\arguments{
\item{x}{dca object created with \code{dca()}}

\item{...}{not used}
}
\value{
a tibble
}
\description{
Export DCA tibble
}
\examples{
dca(cancer ~ cancerpredmarker, data = df_dca) \%>\%
  as_tibble()
}
