% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.dca}
\alias{autoplot.dca}
\title{Export DCA ggplot}
\usage{
\method{autoplot}{dca}(x)
}
\arguments{
\item{x}{dca object created with \code{dca()}}
}
\value{
a ggplot
}
\description{
Export DCA ggplot
}
\examples{
dca(cancer ~ cancerpredmarker, data = df_dca) \%>\%
  autoplot()
}
