% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcurves-package.R
\docType{package}
\name{dcurves-package}
\alias{dcurves}
\alias{dcurves-package}
\title{dcurves: Decision Curve Analysis}
\description{
Diagnostic and prognostic models are typically evaluated with
    measures of accuracy that do not address clinical consequences.
    Decision-analytic techniques allow assessment of clinical outcomes but
    often require collection of additional information may be cumbersome
    to apply to models that yield a continuous result. Decision curve
    analysis is a method for evaluating and comparing prediction models
    that incorporates clinical consequences, requires only the data set on
    which the models are tested, and can be applied to models that have
    either continuous or dichotomous results. The dca function performs
    decision curve analysis for binary outcomes.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ddsjoberg/dcurves/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danieldsjoberg@gmail.com}

}
\keyword{internal}
