% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{df_binary}
\alias{df_binary}
\title{Simulated data with a binary outcome}
\format{
A data frame with 750 rows:
\describe{
\item{patientid}{Identification Number}
\item{cancer}{Cancer Diagnosis: 0=No, 1=Yes}
\item{dead}{Dead (1=yes; 0=no)}
\item{risk_group}{Patient Risk Group (Low, Intermediate, High)}
\item{age}{Patient Age, years}
\item{famhistory}{Family History of Cancer: 0=No, 1=Yes}
\item{marker}{Marker}
\item{cancerpredmarker}{Prob. of Cancer based on Age, Family History, and Marker}
}
}
\usage{
df_binary
}
\description{
Simulated data with a binary outcome
}
\keyword{datasets}
