% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{df_case_control}
\alias{df_case_control}
\title{Simulated data with a case-control outcome}
\format{
A data frame with 750 rows:
\describe{
\item{patientid}{Identification Number}
\item{casecontrol}{Case-control Status: 1=Case, 0=Control}
\item{risk_group}{Patient Risk Group (Low, Intermediate, High)}
\item{age}{Patient Age, years}
\item{famhistory}{Family History of Cancer: 0=No, 1=Yes}
\item{marker}{Marker}
\item{cancerpredmarker}{Prob. of Cancer based on Age, Family History, and Marker}
}
}
\usage{
df_case_control
}
\description{
Simulated data with a case-control outcome
}
\keyword{datasets}
