% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_intervention_above_prev.R
\name{net_intervention_above_prev}
\alias{net_intervention_above_prev}
\title{Net Interventions Avoided Above Prevalence}
\description{
In the software for decision curve analysis, we allow users to specify an
option that shows net benefit in terms of the difference between treat all and treat none.
This, however, does not make sense if threshold probability is greater than the prevalence.
Why ask "what is the net reduction in interventions compared to a treat all
strategy at a threshold where you shouldn’t use the treat all strategy?"
}
