% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_interventions_avoided.R
\name{net_interventions_avoided}
\alias{net_interventions_avoided}
\title{Add Net Interventions Avoided}
\usage{
net_interventions_avoided(x, nper = 100)
}
\arguments{
\item{x}{object of class \code{'dca'} calculated with \code{dca()}}

\item{nper}{Number to report net interventions per. Default is 100.}
}
\value{
'dca' object
}
\description{
Add the number of net interventions avoided to \code{dca()} object.
}
\examples{
dca(cancer ~ cancerpredmarker, data = df_binary) \%>\%
  net_interventions_avoided()
}
