% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dca}
\alias{print.dca}
\title{Print \code{dca()} object}
\usage{
\method{print}{dca}(x, ...)
}
\arguments{
\item{x}{dca object}

\item{...}{not used}
}
\value{
a ggplot
}
\description{
Print \code{dca()} object
}
\examples{
dca(cancer ~ cancerpredmarker, data = df_binary) \%>\%
  print()
}
