% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{as_tibble}
\alias{\%>\%}
\alias{Surv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{survival}{\code{\link[survival]{Surv}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

