% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_net_benefit.R
\name{standardized_net_benefit}
\alias{standardized_net_benefit}
\title{Add Standardized Net Benefit}
\usage{
standardized_net_benefit(x)
}
\arguments{
\item{x}{object of class \code{'dca'} calculated with \code{dca()}}
}
\value{
'dca' object
}
\description{
Add the standardized net benefit to \code{dca()} object.
}
\examples{
dca(Surv(ttcancer, cancer) ~ cancerpredmarker, data = df_surv, time = 1) \%>\%
  standardized_net_benefit()
}
\seealso{
\code{\link[=dca]{dca()}}, \code{\link[=net_intervention_avoided]{net_intervention_avoided()}}, \code{\link[=plot.dca]{plot.dca()}}, \code{\link[=as_tibble.dca]{as_tibble.dca()}}
}
\author{
Daniel D Sjoberg
}
