% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.dca}
\alias{as_tibble.dca}
\title{Convert DCA Object to tibble}
\usage{
\method{as_tibble}{dca}(x, ...)
}
\arguments{
\item{x}{dca object created with \code{dca()}}

\item{...}{not used}
}
\value{
a tibble
}
\description{
Convert DCA Object to tibble
}
\examples{
dca(cancer ~ cancerpredmarker, data = df_binary) \%>\%
  as_tibble()
}
\seealso{
\code{\link[=dca]{dca()}}, \code{\link[=net_intervention_avoided]{net_intervention_avoided()}}, \code{\link[=standardized_net_benefit]{standardized_net_benefit()}}, \code{\link[=plot.dca]{plot.dca()}}
}
\author{
Daniel D Sjoberg
}
