% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.dca}
\alias{autoplot.dca}
\title{Export DCA ggplot}
\usage{
\method{autoplot}{dca}(
  object,
  type = c("net_benefit", "net_intervention_avoided"),
  smooth = FALSE,
  span = 0.2,
  ...
)
}
\arguments{
\item{object}{dca object created with \code{dca()}}

\item{type}{indicates type of plot to produce. must be one of
\code{c("net_benefit", "net_intervention_avoided")}}

\item{smooth}{Logical indicator whether plot will be smooth with
\code{ggplot2::stat_smooth()}. Default is \code{FALSE}}

\item{span}{when \code{smooth = TRUE}, Controls the amount of smoothing for
loess smoother. Smaller numbers produce wigglier lines, larger numbers
produce smoother lines. Default is 0.2.}

\item{...}{not used}
}
\value{
a ggplot
}
\description{
Export DCA ggplot
}
\examples{
dca(cancer ~ cancerpredmarker, data = df_dca) \%>\%
  autoplot()
}
