% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.r
\name{dca}
\alias{dca}
\title{Decision Curve Analysis}
\usage{
dca(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{a data frame containing the outcome of the outcome predictions.}
}
\value{
Returns a list containing the calculated net benefit, ADD MORE
}
\description{
Diagnostic and prognostic models are typically evaluated with measures of accuracy that do not address clinical consequences. Decision-analytic techniques allow assessment of clinical outcomes but often require collection of additional information may be cumbersome to apply to models that yield a continuous result. Decision curve analysis is a method for evaluating and comparing prediction models that incorporates clinical consequences, requires only the data set on which the models are tested, and can be applied to models that have either continuous or dichotomous results. The dca function performs decision curve analysis for binary outcomes.
See http://www.decisioncurveanalysis.org for more information.
}
\examples{
library(MASS)
data.set <- birthwt
model = glm(low ~ age + lwt, family=binomial(link="logit"), data=data.set)
data.set$predlow = predict(model, type="response")
dca(data=data.set, outcome="low", predictors=c("age", "lwt"), probability=c("FALSE", "FALSE"))
result1 = dca(data=data.set, outcome="low", predictors="age", smooth="TRUE", xstop=0.50, probability="FALSE", intervention="TRUE")
result2 = dca(data=data.set, outcome="low", predictors="predlow", smooth="TRUE", xstop=0.50)

}
\author{
Daniel D Sjoberg
}
