% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.r
\name{dca}
\alias{dca}
\title{Perform Decision Curve Analysis}
\usage{
dca(
  formula,
  data,
  thresholds = seq(0.01, 0.99, length.out = 99),
  label = NULL,
  harm = NULL,
  as_probability = character(),
  time = NULL,
  prevalence = NULL
)
}
\arguments{
\item{formula}{formula}

\item{data}{a data frame containing the outcome of the outcome predictions.}

\item{thresholds}{vector of threshold probabilities between 0 and 1.}

\item{label}{named list of variable labels, e.g. \verb{list(age = "Age, years)}}

\item{harm}{named list of harms associated with a test. Default is \code{NULL}}

\item{as_probability}{character vector including names of variables
that will be converted to a probability.}

\item{time}{if outcome is survival, \verb{time=} specifies the time the
assessment is made}

\item{prevalence}{When NULL, the prevalence is estimated from \verb{data=}.
If the data passed is a case-control set, the population prevalence
may be set with this argument.}
}
\value{
List including net benefit of each variable
}
\description{
Diagnostic and prognostic models are typically evaluated with measures of
accuracy that do not address clinical consequences.
Decision-analytic techniques allow assessment of clinical outcomes but often
require collection of additional information may be cumbersome to apply to
models that yield a continuous result. Decision curve analysis is a method
for evaluating and comparing prediction models that incorporates clinical
consequences, requires only the data set on which the models are tested,
and can be applied to models that have either continuous or dichotomous results.
The dca function performs decision curve analysis for binary outcomes.
See http://www.decisioncurveanalysis.org for more information.
}
\examples{
dca(cancer ~ cancerpredmarker, data = df_dca)

dca(Surv(ttcancer, cancer) ~ cancerpredmarker, data = df_dca, time = 1)

}
\author{
Daniel D Sjoberg
}
