% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.r
\name{dca}
\alias{dca}
\title{Perform Decision Curve Analysis}
\usage{
dca(
  formula,
  data,
  thresholds = seq(0.01, 0.99, by = 0.01),
  label = NULL,
  harm = NULL,
  as_probability = character(0L),
  strategies = c("all", "none"),
  time = NULL,
  prevalence = NULL
)
}
\arguments{
\item{formula}{a formula with the outcome on the LHS and a sum of
markers/covariates to test on the RHS}

\item{data}{a data frame containing the variables in \verb{formula=}.}

\item{thresholds}{vector of threshold probabilities between 0 and 1.
Default is \code{seq(0.01, 0.99, by = 0.01)}}

\item{label}{named list of variable labels, e.g. \verb{list(age = "Age, years)}}

\item{harm}{named list of harms associated with a test. Default is \code{NULL}}

\item{as_probability}{character vector including names of variables
that will be converted to a probability.}

\item{strategies}{Default strategies to include: 0, 1, or 2 of ("all", "none").}

\item{time}{if outcome is survival, \verb{time=} specifies the time the
assessment is made}

\item{prevalence}{When \code{NULL}, the prevalence is estimated from \verb{data=}.
If the data passed is a case-control set, the population prevalence
may be set with this argument.}
}
\value{
List including net benefit of each variable
}
\description{
Diagnostic and prognostic models are typically evaluated with measures of
accuracy that do not address clinical consequences.
Decision-analytic techniques allow assessment of clinical outcomes but often
require collection of additional information may be cumbersome to apply to
models that yield a continuous result. Decision curve analysis is a method
for evaluating and comparing prediction models that incorporates clinical
consequences, requires only the data set on which the models are tested,
and can be applied to models that have either continuous or dichotomous
results.
The dca function performs decision curve analysis for binary outcomes.
Review the
\href{http://www.danieldsjoberg.com/dcurves/articles/dca.html}{DCA Vignette}
for a detailed walk-through of various applications.
Also, see http://www.decisioncurveanalysis.org for more information.
}
\examples{
# calculate DCA with binary endpoint
dca(cancer ~ cancerpredmarker + marker,
    data = df_binary,
    as_probability = "marker",
    label = list(cancerpredmarker = "Prediction Model", marker = "Biomarker")) \%>\%
  # plot DCA curves with ggplot
  plot(smooth = TRUE) +
  # add ggplot formatting
  ggplot2::labs(x = "Treatment Threshold Probability")

# calculate DCA with time to event endpoint
dca(Surv(ttcancer, cancer) ~ cancerpredmarker, data = df_surv, time = 1)
}
\seealso{
\code{\link[=net_intervention_avoided]{net_intervention_avoided()}}, \code{\link[=plot.dca]{plot.dca()}},
\code{\link[=as_tibble.dca]{as_tibble.dca()}}
}
\author{
Daniel D Sjoberg
}
