% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcurves-package.R
\docType{package}
\name{dcurves-package}
\alias{dcurves}
\alias{dcurves-package}
\title{dcurves: Decision Curve Analysis for Model Evaluation}
\description{
Diagnostic and prognostic models are typically evaluated with measures of accuracy that do not address clinical consequences. Decision-analytic techniques allow assessment of clinical outcomes, but often require collection of additional information may be cumbersome to apply to models that yield a continuous result. Decision curve analysis is a method for evaluating and comparing prediction models that incorporates clinical consequences, requires only the data set on which the models are tested, and can be applied to models that have either continuous or dichotomous results. See the following references for details on the methods: Vickers (2006) \doi{10.1177/0272989X06295361}, Vickers (2008) \doi{10.1186/1472-6947-8-53}, and Pfeiffer (2020) \doi{10.1002/bimj.201800240}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ddsjoberg/dcurves}
  \item \url{https://www.danieldsjoberg.com/dcurves/}
  \item Report bugs at \url{https://github.com/ddsjoberg/dcurves/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Emily Vertosick \email{vertosie@mskcc.org} [contributor]
}

}
\keyword{internal}
