% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{ggplot}
\alias{aes}
\alias{geom_line}
\alias{stat_smooth}
\alias{coord_cartesian}
\alias{scale_x_continuous}
\alias{labs}
\alias{theme_bw}
\alias{as_tibble}
\alias{\%>\%}
\alias{Surv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{coord_cartesian}}, \code{\link[ggplot2:geom_path]{geom_line}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2:scale_continuous]{scale_x_continuous}}, \code{\link[ggplot2:geom_smooth]{stat_smooth}}, \code{\link[ggplot2:ggtheme]{theme_bw}}}

  \item{survival}{\code{\link[survival]{Surv}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

